Routes are validated and ready to work

+++ controllers/registeredUsers.js
@@ -1,25 +1,66 @@
 const {usersData} = require("../models/UsersData");
+const {validationResult} = require("express-validator")
 
-//GET
+//GET 
 const registeredUsers = async (req, res) => {
-    const registered = await usersData.find(req.body);
-    res.json({registered})
+    try {
+        const protect = validationResult(req)
+        if (protect.isEmpty()) {
+            const registered = await usersData.find(req.body);
+            res.json({registered})
+        } else {
+            res.status(501).json({protect})
+        }
+    } catch (error) {
+        res.status(501).json({error})
+    }
     }
 
 //POST    
 const newUser = async (req, res) => {
-    const users = new usersData (req.body);
-    await users.save();
+    try {
+        const protect = validationResult(req)
+        if (protect.isEmpty()) {
+            const users = new usersData (req.body);
+            await users.save();
+            res.json({registered})
+        } else {
+            res.status(501).json({protect})
+        }
+    } catch (error) {
+        res.status(501).json({error})
+    }
 }
 
 //PUT
 const updateUser = async (req, res) => {
-    await usersData.findByIdAndUpdate(req.params.id, req.body);
+    try {
+        const protect = validationResult(req)
+        if (protect.isEmpty()) {
+            await usersData.findByIdAndUpdate(req.params.id, req.body);
+            res.json({registered})
+        } else {
+            res.status(501).json({protect})
+        }
+    } catch (error) {
+        res.status(501).json({error})
+    }
+    
 }
+
 //DELETE
 const deleteUser =  async (req, res) => {
-    await usersData.findByIdAndDelete(req.params.id);
-
+    try {
+        const protect = validationResult(req)
+        if (protect.isEmpty()) {
+            await usersData.findByIdAndDelete(req.params.id);
+            res.json({registered})
+        } else {
+            res.status(501).json({protect})
+        }
+    } catch (error) {
+        res.status(501).json({error})
+    }
 }
 
 module.exports = {registeredUsers, newUser, updateUser, deleteUser}
